Using one instance of the Gini API SDK and of the single document analyzer
