+++ componentapiexample/src/main/AndroidManifest.xml
@@ -16,6 +16,7 @@
         android:required="true" />
 
     <application
+        android:name="net.gini.android.vision.component.ComponentApiApp"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
