+++ componentapiexample/src/main/java/net/gini/android/vision/component/GiniVisionAppCompatActivity.java
@@ -12,8 +12,6 @@ import android.view.MenuItem;
 import android.view.View;
 import android.widget.Toast;
 
-import net.gini.android.Gini;
-import net.gini.android.SdkBuilder;
 import net.gini.android.models.SpecificExtraction;
 import net.gini.android.vision.Document;
 import net.gini.android.vision.GiniVisionCoordinator;
@@ -55,7 +53,6 @@ public class GiniVisionAppCompatActivity extends AppCompatActivity
 
     private GiniVisionCoordinator mGiniVisionCoordinator;
 
-    private Gini mGiniApi;
     private SingleDocumentAnalyzer mSingleDocumentAnalyzer;
 
     private String mDocumentAnalysisErrorMessage;
@@ -71,8 +68,7 @@ public class GiniVisionAppCompatActivity extends AppCompatActivity
         bindViews();
         setupGiniVisionCoordinator();
         showCamera();
-        createGiniApi();
-        createSingleDocumentAnalyzer();
+        mSingleDocumentAnalyzer = ((ComponentApiApp)getApplication()).getSingleDocumentAnalyzer();
     }
 
     @Override
@@ -84,19 +80,6 @@ public class GiniVisionAppCompatActivity extends AppCompatActivity
             showCamera();
             mShowCameraOnStart = false;
         }
-    }
-
-    private void createGiniApi() {
-        LOG.debug("Instantiating the Gini API with Client ID {}");
-        SdkBuilder builder = new SdkBuilder(this,
-                this.getString(R.string.gini_api_client_id),
-                this.getString(R.string.gini_api_client_secret),
-                "example.com");
-        mGiniApi = builder.build();
-    }
-
-    public void createSingleDocumentAnalyzer() {
-        mSingleDocumentAnalyzer = new SingleDocumentAnalyzer(mGiniApi);
     }
 
     private void setupGiniVisionCoordinator() {
