+++ componentapiexample/src/main/res/values/gini_api_credentials.xml
@@ -1,5 +1,2 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="gini_api_client_id">CLIENT_ID</string>
-    <string name="gini_api_client_secret">CLIENT_SECRET</string>
-</resources>
+<?xml version='1.0'?>
+<resources><string name='gini_api_client_id'>gini-android-test</string><string name='gini_api_client_secret'>H8-PWJYheG6pnTidhZIiCMsGjJg</string></resources>
